add missing package installations in sysfiles

+++ sysfiles/.local/share/sysfiles/makepkg/install.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 # this script configures makepkg to use aria2 for downloading files.
+command -v aria2c >/dev/null || pacman -S --needed aria2 || exit 1
 install -DCvm644 makepkg.extra.conf /etc/makepkg.extra.conf
 f=/etc/makepkg.conf
 [ -e $f ] && ! fgrep -q '. /etc/makepkg.extra.conf' $f &&
